requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pydantic
python-multipart
authlib
httpx